fix: add participationHelper for participation events
