+++ client/src/app/routes/stardust/AddressState.ts
@@ -11,8 +11,8 @@ import { useAliasDetails } from "../../../helpers/hooks/useAliasDetails";
 import { useNftDetails } from "../../../helpers/hooks/useNftDetails";
 import { scrollToTop } from "../../../helpers/pageUtils";
 import { Bech32AddressHelper } from "../../../helpers/stardust/bech32AddressHelper";
+import { deserializeParticipationEventMetadata } from "../../../helpers/stardust/participationHelper";
 import { TransactionsHelper } from "../../../helpers/stardust/transactionsHelper";
-import { deserializeParticipationEventMetadata } from "../../../helpers/stardust/valueFormatHelper";
 import { IBech32AddressDetails } from "../../../models/api/IBech32AddressDetails";
 import { IParticipation } from "../../../models/api/stardust/participation/IParticipation";
 import NetworkContext from "../../context/NetworkContext";
